split and rename
