+++ src/html_utils.php
@@ -101,19 +101,3 @@ function get_required_file_name($fieldName)
     }
     return $fileName;
 }
-
-function str_starts_with(string $haystack, string $needle): bool
-{
-    $length = strlen($needle);
-    return (substr($haystack, 0, $length) === $needle);
-}
-
-function str_ends_with(string $haystack, string $needle): bool
-{
-    $length = strlen($needle);
-    if ($length == 0) {
-        return true;
-    }
-
-    return (substr($haystack, -$length) === $needle);
-}
