+++ src/string_utils.php
@@ -1,0 +1,16 @@
+<?php
+function str_starts_with(string $haystack, string $needle): bool
+{
+    $length = strlen($needle);
+    return (substr($haystack, 0, $length) === $needle);
+}
+
+function str_ends_with(string $haystack, string $needle): bool
+{
+    $length = strlen($needle);
+    if ($length == 0) {
+        return true;
+    }
+
+    return (substr($haystack, -$length) === $needle);
+}
