+++ templates/common/navigation.php
@@ -1,5 +1,5 @@
 <?php
-require_once 'src/form_utils.php';
+require_once 'src/html_utils.php';
 require_once 'src/links.php';
 /**
  * @var array $current
