+++ templates/page/register.php
@@ -1,5 +1,5 @@
 <?php
-require_once 'src/form_utils.php';
+require_once 'src/html_utils.php';
 require_once 'src/links.php';
 ?>
 <form class="form container <?= mark_if_true(!empty($errors), 'form--invalid') ?>"
